+++ src/applications/accredited-representative-portal/representative-form-upload/helpers/index.js
