+++ src/applications/accredited-representative-portal/representative-form-upload/manifest.json
@@ -2,5 +2,5 @@
   "appName": "Representative Form Upload",
   "entryFile": "./app-entry.jsx",
   "entryName": "representative-form-upload",
-  "rootUrl": "/representative-form-upload"
+  "rootUrl": "/representative/representative-form-upload"
 }
