+++ src/applications/accredited-representative-portal/representative-form-upload/package.json
@@ -3,6 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "react-router-dom": "^6.28.0"
+    "react-router-dom": "^6.28.0",
+    "react-router": "3.2.0"
   }
 }
