+++ src/applications/accredited-representative-portal/representative-form-upload/routes.jsx
@@ -5,7 +5,7 @@ import App from './containers/App';
 // Add any new form-upload forms to this list
 const formUploadForms = ['21-686c'];
 
-const config = formConfig();
+const config = formConfig;
 
 const routes = formUploadForms.map(formId => {
   const lowerCaseFormId = formId.toLowerCase();
@@ -19,13 +19,5 @@ const routes = formUploadForms.map(formId => {
     childRoutes: createRoutesWithSaveInProgress(config),
   };
 });
-
-// or dynamic
-// {
-//   path: '/:formId',
-//   component: App,
-//   indexRoute: indexRouteByForm(':formId'),
-//   childRoutes: createRoutesWithSaveInProgress(config),
-// },
 
 export default routes;
