+++ src/applications/accredited-representative-portal/representative-form-upload/tests/unit/helpers/index.unit.spec.js
