--- src/applications/representative-form-upload/app-entry.jsx
@@ -1,13 +1,0 @@
-import 'platform/polyfills';
-import './sass/form-upload.scss';
-
-import startApp from 'platform/startup';
-import routes from './routes';
-import reducer from './reducers';
-import manifest from './manifest.json';
-
-startApp({
-  url: manifest.rootUrl,
-  reducer,
-  routes,
-});
