--- src/applications/representative-form-upload/reducers/index.js
@@ -1,8 +1,0 @@
-import { createSaveInProgressFormReducer } from 'platform/forms/save-in-progress/reducers';
-import formConfig from '../config/form';
-
-const config = formConfig();
-
-export default {
-  form: createSaveInProgressFormReducer(config),
-};
